requirements.txt
discord.py ~= 2.0.0
requests ~= 2.28.1

# development
mypy
flake8
pylint
pytest
setuptools
